Cleanup and refactoring
