+++ screens/LoginScreen.tsx
@@ -1,4 +1,3 @@
-import {StackScreenProps} from '@react-navigation/stack';
 import * as React from 'react';
 import {
   View,
